feat: add toilet font support with bundled fixtures
