+++ tests/fixtures/toilet_future_figlet.txt
@@ -1,0 +1,3 @@
+┏━╸╻┏━╸╻  ┏━╸╺┳╸
+┣╸ ┃┃╺┓┃  ┣╸  ┃ 
+╹  ╹┗━┛┗━╸┗━╸ ╹
