+++ tests/fixtures/toilet_future_hello_rust.txt
@@ -1,0 +1,3 @@
+╻ ╻┏━╸╻  ╻  ┏━┓   ┏━┓╻ ╻┏━┓╺┳╸
+┣━┫┣╸ ┃  ┃  ┃ ┃   ┣┳┛┃ ┃┗━┓ ┃ 
+╹ ╹┗━╸┗━╸┗━╸┗━┛   ╹┗╸┗━┛┗━┛ ╹
