+++ tests/fixtures/toilet_future_negative_float.txt
@@ -1,0 +1,3 @@
+   ╻ ╻ ┏━╸
+╺━╸┗━┫ ┗━┓
+     ╹╹┗━┛
