+++ tests/fixtures/toilet_future_test.txt
@@ -1,0 +1,3 @@
+╺┳╸┏━╸┏━┓╺┳╸
+ ┃ ┣╸ ┗━┓ ┃ 
+ ╹ ┗━╸┗━┛ ╹
