+++ tests/fixtures/toilet_mono12_figlet.txt
@@ -1,0 +1,10 @@
+                                                            
+ ▄▄▄▄▄▄▄▄   ▄▄▄▄▄▄      ▄▄▄▄   ▄▄▄▄                         
+ ██▀▀▀▀▀▀   ▀▀██▀▀    ██▀▀▀▀█  ▀▀██                  ██     
+ ██           ██     ██          ██       ▄████▄   ███████  
+ ███████      ██     ██  ▄▄▄▄    ██      ██▄▄▄▄██    ██     
+ ██           ██     ██  ▀▀██    ██      ██▀▀▀▀▀▀    ██     
+ ██         ▄▄██▄▄    ██▄▄▄██    ██▄▄▄   ▀██▄▄▄▄█    ██▄▄▄  
+ ▀▀         ▀▀▀▀▀▀      ▀▀▀▀      ▀▀▀▀     ▀▀▀▀▀      ▀▀▀▀  
+                                                            
+
