+++ tests/fixtures/toilet_mono12_hello_rust.txt
@@ -1,0 +1,20 @@
+                                                                                
+ ▄▄    ▄▄            ▄▄▄▄      ▄▄▄▄                          ▄▄▄▄▄▄             
+ ██    ██            ▀▀██      ▀▀██                          ██▀▀▀▀██           
+ ██    ██   ▄████▄     ██        ██       ▄████▄             ██    ██  ██    ██ 
+ ████████  ██▄▄▄▄██    ██        ██      ██▀  ▀██            ███████   ██    ██ 
+ ██    ██  ██▀▀▀▀▀▀    ██        ██      ██    ██            ██  ▀██▄  ██    ██ 
+ ██    ██  ▀██▄▄▄▄█    ██▄▄▄     ██▄▄▄   ▀██▄▄██▀            ██    ██  ██▄▄▄███ 
+ ▀▀    ▀▀    ▀▀▀▀▀      ▀▀▀▀      ▀▀▀▀     ▀▀▀▀              ▀▀    ▀▀▀  ▀▀▀▀ ▀▀ 
+                                                                                
+                                                                                
+                                                                                
+                                                                                
+             ██                                                                 
+ ▄▄█████▄  ███████                                                              
+ ██▄▄▄▄ ▀    ██                                                                 
+  ▀▀▀▀██▄    ██                                                                 
+ █▄▄▄▄▄██    ██▄▄▄                                                              
+  ▀▀▀▀▀▀      ▀▀▀▀                                                              
+                                                                                
+
