+++ tests/fixtures/toilet_mono12_negative_float.txt
@@ -1,0 +1,10 @@
+                                        
+               ▄▄▄             ▄▄▄▄▄▄▄  
+              ▄███             ██▀▀▀▀▀  
+             █▀ ██             ██▄▄▄▄   
+           ▄█▀  ██             █▀▀▀▀██▄ 
+  █████    ████████                  ██ 
+                ██      ██     █▄▄▄▄██▀ 
+                ▀▀      ▀▀      ▀▀▀▀▀   
+                                        
+
