+++ tests/fixtures/toilet_mono12_test.txt
@@ -1,0 +1,10 @@
+                                        
+ ▄▄▄▄▄▄▄▄                               
+ ▀▀▀██▀▀▀                        ██     
+    ██      ▄████▄   ▄▄█████▄  ███████  
+    ██     ██▄▄▄▄██  ██▄▄▄▄ ▀    ██     
+    ██     ██▀▀▀▀▀▀   ▀▀▀▀██▄    ██     
+    ██     ▀██▄▄▄▄█  █▄▄▄▄▄██    ██▄▄▄  
+    ▀▀       ▀▀▀▀▀    ▀▀▀▀▀▀      ▀▀▀▀  
+                                        
+
