+++ tests/fixtures/toilet_mono9_figlet.txt
@@ -1,0 +1,8 @@
+                                          
+ ▄▄▄▄▄▄ ▄▄▄▄▄    ▄▄▄  ▀▀█             ▄   
+ █        █    ▄▀   ▀   █     ▄▄▄   ▄▄█▄▄ 
+ █▄▄▄▄▄   █    █   ▄▄   █    █▀  █    █   
+ █        █    █    █   █    █▀▀▀▀    █   
+ █      ▄▄█▄▄   ▀▄▄▄▀   ▀▄▄  ▀█▄▄▀    ▀▄▄ 
+                                          
+
