+++ tests/fixtures/toilet_mono9_hello_rust.txt
@@ -1,0 +1,8 @@
+                                                                      
+ ▄    ▄        ▀▀█    ▀▀█                  ▄▄▄▄▄                  ▄   
+ █    █  ▄▄▄     █      █     ▄▄▄          █   ▀█ ▄   ▄   ▄▄▄   ▄▄█▄▄ 
+ █▄▄▄▄█ █▀  █    █      █    █▀ ▀█         █▄▄▄▄▀ █   █  █   ▀    █   
+ █    █ █▀▀▀▀    █      █    █   █         █   ▀▄ █   █   ▀▀▀▄    █   
+ █    █ ▀█▄▄▀    ▀▄▄    ▀▄▄  ▀█▄█▀         █    ▀ ▀▄▄▀█  ▀▄▄▄▀    ▀▄▄ 
+                                                                      
+
