+++ tests/fixtures/toilet_mono9_negative_float.txt
@@ -1,0 +1,8 @@
+                            
+           ▄▄         ▄▄▄▄▄ 
+          ▄▀█         █     
+         █▀ █         ▀▀▀▀▄▄
+  ▀▀▀   █▄▄▄█▄             █
+            █    █    ▀▄▄▄█▀
+                            
+
