+++ tests/fixtures/toilet_mono9_test.txt
@@ -1,0 +1,8 @@
+                            
+▄▄▄▄▄▄▄                 ▄   
+   █     ▄▄▄    ▄▄▄   ▄▄█▄▄ 
+   █    █▀  █  █   ▀    █   
+   █    █▀▀▀▀   ▀▀▀▄    █   
+   █    ▀█▄▄▀  ▀▄▄▄▀    ▀▄▄ 
+                            
+
