+++ tests/fixtures/toilet_smblock_figlet.txt
@@ -1,0 +1,4 @@
+▛▀▘▜▘▞▀▖▜    ▐  
+▙▄ ▐ ▌▄▖▐ ▞▀▖▜▀ 
+▌  ▐ ▌ ▌▐ ▛▀ ▐ ▖
+▘  ▀▘▝▀  ▘▝▀▘ ▀
