+++ tests/fixtures/toilet_smblock_hello_rust.txt
@@ -1,0 +1,4 @@
+▌ ▌   ▜▜     ▛▀▖      ▐  
+▙▄▌▞▀▖▐▐ ▞▀▖ ▙▄▘▌ ▌▞▀▘▜▀ 
+▌ ▌▛▀ ▐▐ ▌ ▌ ▌▚ ▌ ▌▝▀▖▐ ▖
+▘ ▘▝▀▘ ▘▘▝▀  ▘ ▘▝▀▘▀▀  ▀
