+++ tests/fixtures/toilet_smblock_negative_float.txt
@@ -1,0 +1,4 @@
+   ▌ ▌  ▛▀▘
+▄▄▖▚▄▌  ▙▄ 
+     ▌▗▖▖ ▌
+     ▘▝▘▝▀
