+++ tests/fixtures/toilet_smblock_test.txt
@@ -1,0 +1,4 @@
+▀▛▘     ▐  
+ ▌▞▀▖▞▀▘▜▀ 
+ ▌▛▀ ▝▀▖▐ ▖
+ ▘▝▀▘▀▀  ▀
