+++ tests/fixtures/toilet_wideterm_figlet.txt
@@ -1,0 +1,1 @@
+ＦＩＧｌｅｔ
