+++ tests/fixtures/toilet_wideterm_hello_rust.txt
@@ -1,0 +1,1 @@
+Ｈｅｌｌｏ  Ｒｕｓｔ
