+++ tests/fixtures/toilet_wideterm_negative_float.txt
@@ -1,0 +1,1 @@
+－４．５
