+++ tests/fixtures/toilet_wideterm_test.txt
@@ -1,0 +1,1 @@
+Ｔｅｓｔ
